==> 2-descriptive-outcomes/wasting/0-run-descriptive-outcomes-wasting.R <==

rm(list=ls())
source(paste0(here::here(), "/0-config.R"))


source(here("2-descriptive-outcomes/wasting/1_wast_calcinc.R"))
source(here("2-descriptive-outcomes/wasting/2_birth_wast_outcomes.R"))
source(here("2-descriptive-outcomes/wasting/3_wast_summarize_outcomes.R"))
source(here("2-descriptive-outcomes/wasting/4_co_summarize_outcomes.R"))
source(here("2-descriptive-outcomes/wasting/5_co_flowdata.R"))
source(here("2-descriptive-outcomes/wasting/7_wast_ir_sensitivity.R"))
source(here("2-descriptive-outcomes/wasting/8_wast_manuscript_extra_stats.R"))
source(here("2-descriptive-outcomes/wasting/9_create_wast_seasonality_datasets.R"))
source(here("2-descriptive-outcomes/wasting/10_wast_summarize_outcomes_FE.R"))
source(here("2-descriptive-outcomes/wasting/11_co_summarize_outcomes_FE.R"))
source(here("2-descriptive-outcomes/wasting/12_create_birth_wast_seasonality_dataset.R"))
source(here("2-descriptive-outcomes/wasting/13_birth_Zscore_diff_by_month.R"))
source(here("2-descriptive-outcomes/wasting/14_calc_episode_durations.R"))
source(here("2-descriptive-outcomes/wasting/15_pool_birth_RRs.R"))
source(here("2-descriptive-outcomes/wasting/16_quantify_WLZ_variablity.R"))
source(here("2-descriptive-outcomes/wasting/17_tabulate_co-occurence_categories.R"))
source(here("2-descriptive-outcomes/wasting/18_seasonality_episode_patterns.R"))
source(here("2-descriptive-outcomes/wasting/19_RTM_catch_up_growth.R"))
